refactor: move methods to util module
